Auth and model changes updates

+++ src/main/java/com/api/hotelreviewapplication/dto/ReviewDto.java
@@ -7,5 +7,5 @@ public class ReviewDto {
     private int id;
     private String title;
     private String content;
-    private int stars;
+    private double stars;
 }
